Try setting up project references
