+++ packages/augur-api/package.json
@@ -7,6 +7,7 @@
   "directories": {
     "src": "src"
   },
+  "main": "build",
   "scripts": {
     "version": "yarn rebuild",
     "lint": "tslint -c tslint.json -p . ./src/*.ts ./src/**/*.ts ./src/**/**/*.ts ./src/**/**/**/*.ts",
