+++ packages/augur-api/src/Augur.ts
@@ -4,7 +4,7 @@ import { Contracts } from './api/Contracts';
 import { Trade } from './api/Trade';
 import { GenericAugurInterfaces } from 'augur-core';
 import { NetworkId, NetworkContractAddresses, ContractAddresses } from 'augur-artifacts';
-import * as addressesJson from '../../augur-artifacts/addresses.json';
+import { addresses } from 'augur-artifacts';
 
 export class Augur<TBigNumber> {
   public readonly provider: Provider;
@@ -23,7 +23,7 @@ export class Augur<TBigNumber> {
     this.networkId = networkId;
 
     // API
-    this.networkAddresses = <NetworkContractAddresses>addressesJson;
+    this.networkAddresses = <NetworkContractAddresses>addresses;
     this.addresses = this.networkAddresses[this.networkId];
     this.contracts = new Contracts(this.addresses, this.dependencies);
     this.events = new Events(this.provider, this.addresses.Augur);
