+++ packages/augur-artifacts/package.json
@@ -3,5 +3,6 @@
   "version": "10.0.0-0",
   "description": "ABI and Contract Address Artifacts",
   "author": "Augur Team <team@augur.net>",
-  "license": "MIT"
+  "license": "MIT",
+  "main": "build"
 }
