+++ packages/augur-artifacts/src/contracts.json
