+++ packages/augur-artifacts/src/docker.json
