+++ packages/augur-artifacts/src/upload-block-numbers.json
