+++ packages/augur-state/tsconfig.json
@@ -1,6 +1,16 @@
 {
-  "extends": "../../tsconfig.json",
+  "extends": "../../tsconfig.base.json",
   "compilerOptions": {
-    "typeRoots": ["../augur-types/types"]
-  }
+    "composite": true,
+    "outDir": "./build",
+  },
+  "exclude": ["./src/**/*.test.ts"],
+  "include": ["./src/**/*"],
+  "references": [
+    {"path": "../ethersjs-provider"},
+    {"path": "../augur-core"},
+    {"path": "../augur-types"},
+    {"path": "../augur-api"},
+    {"path": "../augur-artifacts"}
+  ]
 }
