+++ tsconfig.json
@@ -1,68 +1,15 @@
 {
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es2017",
-    "sourceMap": true,
-    "watch": false,
-    "declaration": true,
-    "declarationDir": "definitions",
-    "strictNullChecks": true,
-    "alwaysStrict": true,
-    "noImplicitAny": true,
-    "noImplicitReturns": true,
-    "noImplicitThis": true,
-    "resolveJsonModule": true,
-    "outDir": "build",
-    "baseUrl": ".",
-    "paths": {
-      "augur-core": [
-        "packages/augur-core/source/index.ts"
-      ],
-      "augur-node": [
-        "packages/augur-node/src/index.ts"
-      ],
-      "augur-contract-sdk": [
-        "packages/augur-contract-sdk"
-      ],
-      "augur-tools": [
-        "packages/augur-tools"
-      ],
-      "augur-types": [
-        "packages/augur-types"
-      ],
-      "augur-artifacts": [
-        "packages/augur-artifacts/index.ts"
-      ],
-      "augur-api": [
-        "packages/augur-api/src/index.ts"
-      ],
-      "augur-state": [
-        "packages/augur-state/src/index.ts"
-      ],
-      "contract-dependencies": [
-        "packages/contract-dependencies/src/index.ts"
-      ],
-      "contract-dependencies-ethers": [
-        "packages/contract-dependencies-ethers/src/index.ts"
-      ],
-      "ethers-provider": [
-        "packages/ethersjs-provider/src/index.ts"
-      ]
-    }
-  },
-  "include": [
-    "./node_modules/@types",
-    "./packages",
-    "./packages/augur-core",
-    "./packages/augur-node",
-    "./packages/augur-contract-sdk",
-    "./packages/augur-tools",
-    "./packages/augur-types",
-    "./packages/augur-artifacts",
-    "./packages/augur-api",
-    "./packages/augur-state",
-    "./packages/contract-dependencies",
-    "./packages/contract-dependencies-ethers",
-    "./packages/ethersjs-provider"
+  "files": [],
+  "include": [],
+  "references": [
+    {"path": "packages/ethersjs-provider"},
+    {"path": "packages/augur-core"},
+    {"path": "packages/augur-types"},
+    {"path": "packages/augur-api"},
+    {"path": "packages/augur-node"},
+    {"path": "packages/augur-artifacts"},
+    {"path": "packages/contract-dependencies"},
+    {"path": "packages/augur-state"},
+    {"path": "packages/contract-dependencies-ethers"}
   ]
 }
